DubSync/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
celery==5.3.4

# Audio processing
torch>=2.0.0
torchaudio>=2.0.0
soundfile==0.12.1
librosa==0.10.1
numpy<2.0
scipy==1.11.4

# NeMo and IndicConformer dependencies
nemo-toolkit[asr]==1.22.0
omegaconf==2.3.0
hydra-core==1.3.2
pytorch-lightning==2.1.2
transformers==4.35.2
huggingface-hub==0.19.4

# Additional dependencies
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
psutil==5.9.6
webrtcvad==2.0.10
pydub==0.25.1

# Development
python-decouple==3.8